requirements.txt
blinker==1.6.3
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.5
Flask==3.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonify==0.5
MarkupSafe==2.1.3
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.7.5
six==1.16.0
Werkzeug==3.0.1
